to store the messages as keys and their corresponding timestamps as values.
Whenever a new message is received, we check if it already exists in the dictionary and if the time difference between the current timestamp and the stored timestamp is less than 10 seconds. If it is, we return False to indicate that the message should not be printed. Otherwise, we update the timestamp in the dictionary and return True to indicate that the message should be printed.

The dictionary data structure is used here to efficiently store and retrieve the messages and their timestamps. By using the message as the key, we can quickly check if a message has been logged before and retrieve its timestamp in constant time.